+++ core/src/main/scala/cromwell/core/WorkflowMetadataKeys.scala
@@ -35,6 +35,4 @@ object WorkflowMetadataKeys {
   val SummaryNameDecreasing = "WORKFLOW_METADATA_SUMMARY_ENTRY_DECREASING"
 
   val Labels = "labels"
-
-  val MetadataSource = "metadataSource"
 }