+++ engine/src/main/scala/cromwell/engine/io/gcs/GcsBatchFlow.scala
@@ -1,7 +1,8 @@
 package cromwell.engine.io.gcs
 
-import java.io.IOException
+import akka.NotUsed
 
+import java.io.IOException
 import akka.actor.Scheduler
 import akka.stream._
 import akka.stream.scaladsl.{Flow, GraphDSL, MergePreferred, Partition}
@@ -13,10 +14,11 @@ import com.typesafe.scalalogging.StrictLogging
 import common.util.StringUtil.EnhancedToStringable
 import cromwell.cloudsupport.gcp.GoogleConfiguration
 import cromwell.cloudsupport.gcp.gcs.GcsStorage
+import cromwell.core.io.IoAck
 import cromwell.engine.io.IoActor._
 import cromwell.engine.io.IoAttempts.EnhancedCromwellIoException
-import cromwell.engine.io.RetryableRequestSupport.{isRetryable, isInfinitelyRetryable}
-import cromwell.engine.io.gcs.GcsBatchFlow.{BatchFailedException, _}
+import cromwell.engine.io.RetryableRequestSupport.{isInfinitelyRetryable, isRetryable}
+import cromwell.engine.io.gcs.GcsBatchFlow._
 import cromwell.engine.io.{IoAttempts, IoCommandContext}
 import mouse.boolean._
 
@@ -54,17 +56,49 @@ class GcsBatchFlow(batchSize: Int, scheduler: Scheduler, onRetry: IoCommandConte
     }
   }
 
-  private val batchRequest: BatchRequest = {
-    val storage = new Storage.Builder(
+  /**
+    * Returns a new BatchRequest instance.
+    *
+    * The correlated errors that were seen were:
+    * - Via CROM-6708 timeouts were occurring, followed usually by...
+    * - Via CROM-6709 batches had grown beyond 3001+ requests.
+    *
+    * Examining BatchRequest.execute(), if any exception occurs during the method, either during the HTTP request or
+    * during one of the response handlers, the BatchRequest.execute() method simply throws the exception.
+    *
+    * When exceptions are NOT thrown the BatchRequest's internal queue is only partially or fully cleared:
+    * - partial : https://github.com/googleapis/google-api-java-client/blob/v1.31.1/google-api-client/src/main/java/com/google/api/client/googleapis/batch/BatchRequest.java#L275
+    * -   full  : https://github.com/googleapis/google-api-java-client/blob/v1.31.1/google-api-client/src/main/java/com/google/api/client/googleapis/batch/BatchRequest.java#L281
+    *
+    * The BatchRequest's internal queue is NOT cleared when exceptions are thrown inside BatchRequest.execute().
+    * Any subsequent enqueueing operations only add to the internal queue and do not replace the internal queue.
+    * So eventually timeouts lead to large batches of 1000+ elements that cause further timeout exceptions too.
+    * Those failed batch elements were re-appended leading to batches of greater than 3000 requests.
+    * At least that's the theory.
+    *
+    * Instead we'll create a new BatchRequest each time we run GcsBatchFlow.executeBatch().
+    *
+    * From basic performance testing with YourKit we were able to create 1,000,000 of these BatchRequest objects in
+    * under 17s. However, if needed a cached var batchRequest instance could be implemented that only recreates the
+    * object when the internal queue is "dirty" with a .size() > 0.
+    */
+  private def newBatchRequest(): BatchRequest = {
+    val builder = new Storage.Builder(
       GcsStorage.HttpTransport,
       JacksonFactory.getDefaultInstance,
       httpRequestInitializer
     ).setApplicationName(applicationName)
-
-    storage.build().batch()
+    val client = builder.build()
+    client.batch(client.getRequestFactory.getInitializer)
   }
 
-  val flow = GraphDSL.create() { implicit builder =>
+  val flow: Graph[
+    FlowShape[
+      GcsBatchCommandContext[_, _],
+      (IoAck[_], IoCommandContext[_]),
+    ],
+    NotUsed,
+  ] = GraphDSL.create() { implicit builder =>
     import GraphDSL.Implicits._
 
     // Source where batch commands are coming from. This is the input port of this flow
@@ -122,10 +156,12 @@ class GcsBatchFlow(batchSize: Int, scheduler: Scheduler, onRetry: IoCommandConte
   }
 
   private def executeBatch(contexts: Seq[GcsBatchCommandContext[_, _]]): List[Future[GcsBatchResponse[_]]] = {
-    def failAllPromisesWith(failure: Throwable) = contexts foreach { context =>
+    def failAllPromisesWith(failure: Throwable): Unit = contexts foreach { context =>
       context.promise.tryFailure(failure)
       ()
     }
+
+    val batchRequest = newBatchRequest()
 
     // Add all requests to the batch
     contexts foreach { _.queue(batchRequest) }
@@ -136,12 +172,12 @@ class GcsBatchFlow(batchSize: Int, scheduler: Scheduler, onRetry: IoCommandConte
     // Otherwise fail with the original exception
     Try(batchRequest.execute()) match {
       case Failure(failure: IOException) =>
-        logger.info(s"Failed to execute GCS Batch request. Failed request belonged to batch of size ${batchCommandNamesList.size} containing commands: " +
-          s"${batchCommandNamesList.mkString("\n")}.", failure.toPrettyElidedString(limit = 1000))
+        logger.info(s"Failed to execute GCS Batch request. Failed request belonged to batch of size ${batchRequest.size()} containing commands: " +
+          s"${batchCommandNamesList.mkString("\n")}.\n${failure.toPrettyElidedString(limit = 1000)}")
         failAllPromisesWith(BatchFailedException(failure))
       case Failure(failure) =>
-        logger.info(s"Failed to execute GCS Batch request. Failed request belonged to batch of size ${batchCommandNamesList.size} containing commands: " +
-          s"${batchCommandNamesList.mkString("\n")}.", failure.toPrettyElidedString(limit = 1000))
+        logger.info(s"Failed to execute GCS Batch request. Failed request belonged to batch of size ${batchRequest.size()} containing commands: " +
+          s"${batchCommandNamesList.mkString("\n")}.\n${failure.toPrettyElidedString(limit = 1000)}")
         failAllPromisesWith(failure)
       case _ =>
     }